move files to lib and change relative folder
