+++ lib/fs.js
